# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de Atendimento",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/funcionarios/find/{funcionario}": {
            "get": {
                "tags": [
                    "/funcionarios/find/:funcionario"
                ],
                "summary": "Pega funcionario por Id",
                "description": "Retorna funcionario por Id",
                "operationId": "App\\Http\\Controllers\\FuncionarioController::show",
                "parameters": [
                    {
                        "name": "funcionario",
                        "in": "path",
                        "description": "funcionario id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/funcionarios/create": {
            "post": {
                "tags": [
                    "/funcionarios/create"
                ],
                "summary": "Cria um novo funcionario",
                "description": "Retorna um novo funcionario",
                "operationId": "App\\Http\\Controllers\\FuncionarioController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/funcionarios/update/{funcionario}": {
            "put": {
                "tags": [
                    "/funcionarios/update/:funcionario"
                ],
                "summary": "Atualiza funcionario existente",
                "description": "Retorna funcionario atualizado por id",
                "operationId": "App\\Http\\Controllers\\FuncionarioController::update",
                "parameters": [
                    {
                        "name": "funcionario",
                        "in": "path",
                        "description": "funcionario id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "servico_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Sucesso"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/funcionarios/{funcionario}": {
            "delete": {
                "tags": [
                    "/funcionarios/:funcionario"
                ],
                "summary": "Deleta um funcionario existente",
                "description": "Deleta o funcionario por Id",
                "operationId": "App\\Http\\Controllers\\FuncionarioController::delete",
                "parameters": [
                    {
                        "name": "funcionario",
                        "in": "path",
                        "description": "funcionario id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}